agents/grocery_optimizer/requirements.txt
msal==1.31.0
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
pyyaml==6.0.2
